Move build logic to RunListener extension
Instead of extending BuildWrapper for pre-build actions and BuildStep or
Notifier/Recorder for post-build actions, use RunListener for boths
actions.
The main benefit of this is that there is no need for user to select
pre-build and post-build actions in the GUI.
Another advantage of using RunListener and its method onCompleted to run
post-build actions is that at the time onCompleted is called, the build
is already finished, together with all its steps, and this we have a
definite result of the build.

The disadvantage on the other hand is that RunListener is called
hudson-wise for every build using jenkins, even if its running on node
which has no actions configured. This is not really optimal.